Python Position Analysis
Trying to write particle tracker in Python, I passed in the bounding box positions and the minor axis of the fitted ellipse through a .txt file. 
As of now, it can scatterplot the data, I'm trying to make an animation or something that shows the movement of the cells first.